openetelemetry/prometheus_grafana_flask/requirements.txt
blinker==1.7.0
click==8.1.7
Deprecated==1.2.14
Flask==3.0.2
importlib-metadata==6.11.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
opentelemetry-api==1.22.0
opentelemetry-distro==0.43b0
opentelemetry-exporter-prometheus==0.43b0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-aws-lambda==0.43b0
opentelemetry-instrumentation-dbapi==0.43b0
opentelemetry-instrumentation-jinja2==0.43b0
opentelemetry-instrumentation-logging==0.43b0
opentelemetry-instrumentation-sqlite3==0.43b0
opentelemetry-instrumentation-urllib==0.43b0
opentelemetry-instrumentation-wsgi==0.43b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
prometheus-client==0.19.0
typing_extensions==4.9.0
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
